+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- double builtin added:
+  `double`, `flintmax`, `realmax`
+
 - assert_functions builtin added:
   `assert_istrue`, `assert`
   `assert_isfalse`, `assert_checkerror`